Added Interactive Communication Files
Here, A project was created using Arduino IDE and the Adafruit-Flora for individuals with disabilities. The person can use buttons housed within custom laser-printed boxes to move a cursor up, down, left and right on the screen, as well as click. LEDs can be used for direct purpose of use for the project (such as where the cursor needs to be before clicking) and/or light up to show when to click. Alternatively, sounds can be played.